Organize folders and methods
