+++ src/index.html
@@ -32,9 +32,9 @@
         </form>
       </div>
     </nav>
-    <div class="">
-      <h1 id="scoreBox">Hello</h1>
-    </div>
+    <div>
+    <div id="scoreBox">Hello</div>
     <div id="canvas"></div>
+  </div>
   </body>
 </html>
